test_dummy.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Dummy Placeholder</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/">Dummy Target</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

function iterateMethods(label, root, suppress) {
    for (var f in root) {
        if (suppress && suppress.indexOf(f) != -1)
            continue;
        if (typeof root[f] === 'function')
            arr.push(label + f);
    }
}
var arr = [];
var enum_list_f = ['getInstalled','getSelf','install','mgmt.addEventListener','mgmt.getAll','mgmt.removeEventListener']

enum_list_f = enum_list_f.map(function (x) { return "navigator.mozApps." + x;});

iterateMethods("navigator.mozApps.", navigator.mozApps);
iterateMethods("navigator.mozApps.mgmt.", navigator.mozApps.mgmt);

arr = arr.sort();

for (var i = 0; i < arr.length; i++) {
    is(arr[i], enum_list_f[i], "Unexpected exposed element " + arr[i] + "was expecting " + enum_list_f[i])
}

</script>
</pre>
</body>
</html>
